// src/types/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.4"
  }
  public: {
    Tables: {
// ...
      products: {
// ...
        Insert: {
          additional_terms?: string | null
          application?: string | null
          category_id?: string | null
          cleaning_info?: string | null
          components_determining_hazard?: string | null
          composants_determinant_le_danger?: string | null
          conditions_of_sale?: string | null
          conseils_de_prudence?: string | null
          consignes_de_stockage?: string | null
          consignes_delimination?: string | null
          coverage?: string | null
          created_at?: string | null
          description?: string | null
          disposal?: string | null
          do_not_freeze?: boolean | null
          emergency_response_guide?: string | null
          features?: string | null
          first_aid?: string | null
          green_conscious?: boolean | null
          hazard_class?: Database["public"]["Enums"]["hazard_class"] | null
          hazard_statements?: string | null
          id?: string
          inherent_risk?: string | null
          is_active?: boolean | null
          limitations?: string | null
          manufacturing_notice?: string | null
          mentions_de_danger?: string | null
          mesures_de_premiers_secours?: string | null
          mix_well?: boolean | null
          mot_de_signalement?: string | null
          name: string
          packing_group?: Database["public"]["Enums"]["packing_group"] | null
          pictogram_urls?: string | null
          pictograms?: string | null
          precautionary_statements?: string | null
          premiers_soins?: string | null
          proper_shipping_name?: string | null
          response_statements?: string | null
          safety_notice?: string | null
          shelf_life?: string | null
          short_description_english?: string | null
          short_description_french?: string | null
          short_description_spanish?: string | null
          signal_word?: Database["public"]["Enums"]["hazard_signal"] | null
          sku?: string | null
          slug: string
          sort_order?: number | null
          storage?: string | null
          subtitle_1?: string | null
          subtitle_2?: string | null
          test_data?: string | null
          un_number?: string | null
          updated_at?: string | null
          used_by_date?: string | null
          voc_data?: string | null
          warranty_limitation?: string | null
        }
        Update: {
          additional_terms?: string | null
          application?: string | null
          category_id?: string | null
          cleaning_info?: string | null
          components_determining_hazard?: string | null
          composants_determinant_le_danger?: string | null
          conditions_of_sale?: string | null
          conseils_de_prudence?: string | null
          consignes_de_stockage?: string | null
          consignes_delimination?: string | null
          coverage?: string | null
          created_at?: string | null
          description?: string | null
          disposal?: string | null
          do_not_freeze?: boolean | null
          emergency_response_guide?: string | null
          features?: string | null
          first_aid?: string | null
          green_conscious?: boolean | null
          hazard_class?: Database["public"]["Enums"]["hazard_class"] | null
          hazard_statements?: string | null
          id?: string
          inherent_risk?: string | null
          is_active?: boolean | null
          limitations?: string | null
          manufacturing_notice?: string | null
          mentions_de_danger?: string | null
          mesures_de_premiers_secours?: string | null
          mix_well?: boolean | null
          mot_de_signalement?: string | null
          name?: string
          packing_group?: Database["public"]["Enums"]["packing_group"] | null
          pictogram_urls?: string | null
          pictograms?: string | null
          precautionary_statements?: string | null
          premiers_soins?: string | null
          proper_shipping_name?: string | null
          response_statements?: string | null
          safety_notice?: string | null
          shelf_life?: string | null
          short_description_english?: string | null
          short_description_french?: string | null
          short_description_spanish?: string | null
          signal_word?: Database["public"]["Enums"]["hazard_signal"] | null
          sku?: string | null
          slug?: string
          sort_order?: number | null
          storage?: string | null
          subtitle_1?: string | null
          subtitle_2?: string | null
          test_data?: string | null
          un_number?: string | null
          updated_at?: string | null
          used_by_date?: string | null
          voc_data?: string | null
          warranty_limitation?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "products_category_id_fkey"
            columns: ["category_id"]
            isOneToOne: false
            referencedRelation: "categories"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never